requirements.txt
openai>=1.0.0
openai-whisper
python-dotenv
